Move method to get service dialog name into rest/gen_data
remove unecessary enumerate in test_rest_services
